% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design}
\alias{design}
\title{Design generator.}
\usage{
design(n, k, method, l = 0, u = 1, as.df = TRUE, ...)
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
Number of design points (rows).}

\item{k}{[\code{integer(1)}]\cr
Number of variables (columns).}

\item{method}{[\code{character(1)}]\cr
Name of method used to generate the design. Possible values are
\describe{
  \item{uniform}{Uniform random sampling.}
  \item{improvedlhs}{Delegate to \code{\link[lhs]{improvedLHS}}.}
  \item{maximinlhs}{Delegate to \code{\link[lhs]{maximinLHS}}.}
  \item{geneticlhs}{Delegate to \code{\link[lhs]{geneticLHS}}.}
  \item{halton}{Delegate to \code{\link[randtoolbox]{halton}}.}
  \item{sobol}{Delegate to \code{\link[randtoolbox]{sobol}} with option scrambling=3.}
}}

\item{l}{[\code{numeric}]\cr
Lower bound for design points. Either a single numeric value
or a vector of length \code{k}.
Default is 0.}

\item{u}{[\code{numeric}]\cr
Upper bound for design points. Either a single numeric value
or a vector of length \code{k}.
Default is 1.}

\item{as.df}{[\code{logical(1)}]\cr
Return points as data frame?
Default is \code{TRUE}.}

\item{...}{[any]\cr
Further parameters passed down to generator if applicable.}
}
\value{
Design points as a matrix or data.frame (see parameter \code{as.df}).
}
\description{
Generate pseudo-random (uniform, latin-hypercube samle) or quasi-random
(Halton and Sobol sequences) designs. Basically a wrapper
around different functions from packages \pkg{lhs} and \pkg{randtoolbox}.
}
\examples{
methods = getSupportedMethods()
designs = lapply(methods, function(method) {
  design(n = 100, k = 2, method = method, l = -5, u = c(5, 10), as.df = FALSE)
})

# pass down options to generator
d = design(n = 50, k = 4, method = "sobol", scrambling = 2, seed = 123)
print(d)
}
