% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{getSupportedMethods}
\alias{getSupportedMethods}
\title{Get character vector of supported design generators,
i.e., possible value for parameter \code{method} of
function \code{\link{design}}.}
\usage{
getSupportedMethods()
}
\description{
Get character vector of supported design generators,
i.e., possible value for parameter \code{method} of
function \code{\link{design}}.
}
