% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stardiscrepancy.R
\name{stardiscrepancy}
\alias{stardiscrepancy}
\title{Calculate star discrepancy of a set of points.}
\usage{
stardiscrepancy(x)
}
\arguments{
\item{x}{[\code{matrix(n, d)}]\cr
An \eqn{n \times d} matrix where \eqn{n} is the number of points and \eqn{d}
is the dimension of the search space.}
}
\value{
[\numeric(1)]\cr Star discrepancy of \code{x}.
}
\description{
To be written ...
}
