% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{getSupportedMethods}
\alias{getSupportedMethods}
\title{Get character vector of supported design generators.}
\usage{
getSupportedMethods()
}
\description{
Get character vector of supported design generators.
}
